requirements.txt
pyyaml>=6.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
shap>=0.41.0
plotly>=5.10.0
dash>=2.6.0
mlflow>=2.0.0
delta-spark>=2.2.0
pyspark>=3.4.0